Move issues to top level of src/
With the schema XML loading code (the other original part of the "common"
folder) consolidated with the rest of the schema code, having two
folder levels for the issues is unnecessary.
