+++ spec_tests/jsonTests.spec.js
@@ -7,7 +7,7 @@ import { buildSchemas } from '../src/schema/init'
 import { SchemaSpec, SchemasSpec } from '../src/schema/specs'
 import path from 'path'
 import { BidsSidecar, BidsTsvFile } from '../src/bids'
-import { generateIssue, IssueError } from '../src/common/issues/issues'
+import { generateIssue, IssueError } from '../src/issues/issues'
 import { DefinitionManager } from '../src/parser/definitionManager'
 import parseTSV from '../src/bids/tsvParser'
 import { shouldRun } from '../tests/testUtilities'
