+++ src/bids/schema.js
@@ -2,7 +2,7 @@ import castArray from 'lodash/castArray'
 import semver from 'semver'
 
 import { buildSchemas } from '../schema/init'
-import { IssueError } from '../common/issues/issues'
+import { IssueError } from '../issues/issues'
 import { SchemaSpec, SchemasSpec } from '../schema/specs'
 
 const alphabeticRegExp = new RegExp('^[a-zA-Z]+$')
