+++ src/bids/validator/tsvValidator.js
@@ -3,7 +3,7 @@ import { BidsTsvElement, BidsTsvRow } from '../types/tsv'
 import { BidsValidator } from './validator'
 import { parseHedString } from '../../parser/parser'
 import ParsedHedString from '../../parser/parsedHedString'
-import { generateIssue } from '../../common/issues/issues'
+import { generateIssue } from '../../issues/issues'
 import { ReservedChecker } from '../../parser/reservedChecker'
 import { getTagListString } from '../../parser/parseUtils'
 import { EventManager } from '../../parser/eventManager'
