+++ tests/testData/stringParserTests.data.js
@@ -1,4 +1,4 @@
-import { generateIssue } from '../../src/common/issues/issues'
+import { generateIssue } from '../../src/issues/issues'
 
 export const parseTestData = [
   {
